Fnc : Rsynupdate : Possibilité de choisir en mode interactif les sites a mettre a jour depuis le fichier rsyncupdate.conf
Fnc : Ajout de l'envoi de mails en cas d'alertes pour le statut de redirection Apache, le statut de réplication SQl et la latence de réplication SQL

Ref : Correction du format de la date écrites dans svnstatus.txt du script update.php

git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@15482 a21d6be2-e512-0410-a060-91f5a2946631

+++ shell/rsyncupdate.sh
@@ -41,17 +41,20 @@ then
     # Skip comment lines and empty lines
     if [ "$first_character" != "#" ] && [ "$first_character" != "" ]
     then
-      echo "-- Rsync $line --"
-      rsync -avpz --stats $BASH_PATH/.. --delete $line --exclude-from=$BASH_PATH/rsyncupdate.exclude \
-        --exclude includes/config_overload.php \
-        --exclude tmp \
-        --exclude lib \
-        --exclude files \
-        --exclude includes/config.php \
-        --exclude images/pictures/logo_custom.png
-      check_errs $? "Failed to rsync $line" "Succesfully rsync-ed $line"
-      scp $BASH_PATH/../tmp/svnlog.txt $line/tmp/svnlog.txt
-      scp $BASH_PATH/../tmp/svnstatus.txt $line/tmp/svnstatus.txt
+      echo "Do you want to update $line (y or n) [default n] ? \c" ; read REPLY < /dev/tty
+      if [ "$REPLY" = "y" ] ; then
+        echo "-- Rsync $line --"
+        rsync -avpz --stats $BASH_PATH/.. --delete $line --exclude-from=$BASH_PATH/rsyncupdate.exclude \
+          --exclude includes/config_overload.php \
+          --exclude tmp \
+          --exclude lib \
+          --exclude files \
+          --exclude includes/config.php \
+          --exclude images/pictures/logo_custom.png
+        check_errs $? "Failed to rsync $line" "Succesfully rsync-ed $line"
+        scp $BASH_PATH/../tmp/svnlog.txt $line/tmp/svnlog.txt
+        scp $BASH_PATH/../tmp/svnstatus.txt $line/tmp/svnstatus.txt
+      fi
     fi
   done < $BASH_PATH/rsyncupdate.conf
